Extract DiscountCalculatorForm and its types
